move client to its own submodule
